remove wait logic in operator
